statether/x/statether/types/message_remove_address_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"statether/testutil/sample"
)

func TestMsgRemoveAddress_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgRemoveAddress
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgRemoveAddress{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgRemoveAddress{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
